requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.2
--find-links=https://download.pytorch.org/whl/torch_stable.html
torch==2.0.1+cpu
torchvision==0.15.2+cpu
timm==0.9.2